spark_core/requirements.txt
fastapi
uvicorn
websockets
httpx
psutil
PyJWT
pydantic
